+++ code/include/solver.hpp
@@ -65,8 +65,10 @@ class Solver {
     // Returns whether at least one action was eliminated
     [[nodiscard]]
     auto prep_dominated_actions_(std::vector<std::vector<unsigned int>>& landmarks) -> bool;
-    void prep_orbital_probing_();
-    void prep_initial_action_sequencing_(std::vector<std::vector<unsigned int>>& landmarks);
+    [[nodiscard]]
+    auto prep_orbital_probing_() -> bool;
+    [[nodiscard]]
+    auto prep_initial_action_sequencing_(std::vector<std::vector<unsigned int>>& landmarks) -> bool;
     void prep_eliminated_facts_(std::vector<std::vector<unsigned int>>& landmarks);
     void prep_eliminated_actions_();
     void prep_fixed_actions_();